+++ CHANGELOG.md
@@ -6,10 +6,11 @@ FEATURES:
 
 ENHANCEMENTS:
 * mdb: support `security_group_ids` for managed service for kafka
-
-ENHANCEMENTS:
 * mdb: add `format_schema` section in `yandex_mdb_clickhouse_cluster` resource and data source
 * mdb: add `ml_model` section in `yandex_mdb_clickhouse_cluster` resource and data source
+* mdb: add `sql_user_management` and `sql_database_management` attributes in `yandex_mdb_clickhouse_cluster` resource and data_source
+* mdb: add `admin_password` attribute in `yandex_mdb_clickhouse_cluster` resource
+
 
 ## 0.49.0 (January 15, 2021)
 FEATURES: